rpk: Add 'rpk redpanda mode', deprecate 'rpk mode'
